Prevent re-reacting to messages that are not in the DB
Previously if a batch of messages was missing from the database
(this can happen if the program was e.g. run on different
machines that do not share the database), all of the previously
received (and reacted) messages would get a new response, which
might in the long run have catastrophic effects.

After this commit, when reading messages from Habitica API, if
there is a more recent message sent to the sender of a given message,
the newly-read message will not be marked as reaction_pending. This
works as long as the API returns the messages sorted from newest to
oldest, which seems to be the case currently.