Split long private messages
When a message that is longer than the Habitica maximum message length
of 3000 characters is sent, it is split into separate messages at
newlines. If the message cannot be split (a sequence of more than 3000
characters without newlines) or the message is overly long (taking more
than three messages to send) an error is raised and no message is sent.