Update wiki parsing after Wikia update
Wikia updated the site, so the old parsing logic was no longer able to
identify the content from the wiki page. Updated the code and wget'ed
the current page to the test data. This also required some minor changes
to the tests, as the expected values parsed from the page changed.