Sen a PM when removing a member from the party due to inactivity
Now the same message is both broadcasted via the actual party member
removal and sent as a PM to the member. This is due to it being hard to
verify that the message provided to the API has actually been shown to
the user, whereas a PM will leave a "paper trail".